Performance tests improvements (#1270)
* Refs #7684. Large data support in performance tests.


* Refs #7685. Fix multiple subscribers in latency test.


* Refs #7924. Demands file for latency test.


* Refs 7923. Refactor of csv generation & support for csv for any sample size.


* Refs #7685. Multiple subscribers in throughput test.


* Refs #7685. Multiple subscribers intraprocess in latency test.


* Refs #7685. Suggested changes applied.


* Refs #7685. UDP transport descriptors added to XML files.


* Refs #7685. Fix latency test huge complete time when samples lost.


* Refs #7685. Suggestions applied.


Refs #7685. Suggestions applied.


* Refs #7685. LatencyTest default sample sizes changed.


* Refs #7685. Move XML <domain_id> field.


* Refs #7685. Max demand from 1MB => 256KB because it takes too long in CI tests.


* Refs #7685. Latency test cmake & python demands_file support.


* Refs #7685. Bugfix ThroughputTest: Publisher waits indefinitely when subscriber unmatch is not received.


* Refs #7685. Set old demands sizes for both tests.


* Refs 8798: Add intraprocess performance test cases


* Refs #8798: Fix throughput test closing for intraprocess


* Refs #7971: Remove development test


* Refs #8798: Demands file for latency intraprocess cases


* Refs #8798: Support for large data in Throughput test


* Refs #8798: Solve conversion warning on Windows


* Refs #8798: Uncrustify


* Refs #8798: Uncrustify 0.71


* Refs #8798: Uncrustify ThroughputTypes.hpp


* Refs #8798: Apply suggestions

+++ test/performance/throughput/ThroughputTypes.cpp
@@ -24,8 +24,10 @@
 using namespace eprosima::fastrtps;
 using namespace eprosima::fastrtps::rtps;
 
-//Funciones de serializacion y deserializacion para el ejemplo
-bool ThroughputDataType::serialize(void*data,SerializedPayload_t* payload)
+// Serialization and deserialization functions
+bool ThroughputDataType::serialize(
+        void* data,
+        SerializedPayload_t* payload)
 {
     ThroughputType* lt = (ThroughputType*)data;
     memcpy(payload->data, &lt->seqnum, sizeof(lt->seqnum));
@@ -33,13 +35,15 @@ bool ThroughputDataType::serialize(void*data,SerializedPayload_t* payload)
     memcpy(payload->data + 4, &data_size, sizeof(data_size));
     //	std::copy(lt->data.begin(),lt->data.end(),payload->data+8);
     memcpy(payload->data + 8, lt->data.data(), lt->data.size());
-    payload->length = 8+(uint16_t)lt->data.size();
+    payload->length = 8 + static_cast<uint32_t>(lt->data.size());
     return true;
 }
 
-bool ThroughputDataType::deserialize(SerializedPayload_t* payload,void * data)
+bool ThroughputDataType::deserialize(
+        SerializedPayload_t* payload,
+        void* data)
 {
-    if(payload->length > 0)
+    if (payload->length > 0)
     {
         ThroughputType* lt = (ThroughputType*)data;
         memcpy(&lt->seqnum, payload->data, sizeof(lt->seqnum));
@@ -53,31 +57,34 @@ bool ThroughputDataType::deserialize(SerializedPayload_t* payload,void * data)
     return true;
 }
 
-std::function<uint32_t()> ThroughputDataType::getSerializedSizeProvider(void* data)
+std::function<uint32_t()> ThroughputDataType::getSerializedSizeProvider(
+        void* data)
 {
     return [data]() -> uint32_t
-    {
-        ThroughputType *tdata = static_cast<ThroughputType*>(data);
-        uint32_t size = 0;
+           {
+               ThroughputType* tdata = static_cast<ThroughputType*>(data);
+               uint32_t size = 0;
 
-        size = (uint32_t)(sizeof(uint32_t) + sizeof(uint32_t) + tdata->data.size());
+               size = (uint32_t)(sizeof(uint32_t) + sizeof(uint32_t) + tdata->data.size());
 
-        return size;
-    };
+               return size;
+           };
 }
 
 void* ThroughputDataType::createData()
 {
-    return (void*)new ThroughputType((uint16_t)this->m_typeSize);
+    return (void*)new ThroughputType(this->m_typeSize);
 }
-void ThroughputDataType::deleteData(void* data)
+
+void ThroughputDataType::deleteData(
+        void* data)
 {
     delete((ThroughputType*)data);
 }
 
-
-
-bool ThroughputCommandDataType::serialize(void*data,SerializedPayload_t* p)
+bool ThroughputCommandDataType::serialize(
+        void* data,
+        SerializedPayload_t* p)
 {
     ThroughputCommandType* t = (ThroughputCommandType*)data;
     p->length = 0;
@@ -96,7 +103,10 @@ bool ThroughputCommandDataType::serialize(void*data,SerializedPayload_t* p)
     p->length += sizeof(t->m_totaltime);
     return true;
 }
-bool ThroughputCommandDataType::deserialize(SerializedPayload_t* p,void * data)
+
+bool ThroughputCommandDataType::deserialize(
+        SerializedPayload_t* p,
+        void* data)
 {
     ThroughputCommandType* t = (ThroughputCommandType*)data;
     p->pos = 0;
@@ -117,24 +127,27 @@ bool ThroughputCommandDataType::deserialize(SerializedPayload_t* p,void * data)
     return true;
 }
 
-std::function<uint32_t()> ThroughputCommandDataType::getSerializedSizeProvider(void*)
+std::function<uint32_t()> ThroughputCommandDataType::getSerializedSizeProvider(
+        void*)
 {
     return []() -> uint32_t
-    {
-        uint32_t size = 0;
+           {
+               uint32_t size = 0;
 
-        size = (uint32_t)(sizeof(uint32_t) + sizeof(uint32_t) + sizeof(uint32_t)  + sizeof(uint32_t) +
-                sizeof(uint64_t) + sizeof(uint64_t));
+               size = (uint32_t)(sizeof(uint32_t) + sizeof(uint32_t) + sizeof(uint32_t)  + sizeof(uint32_t) +
+                       sizeof(uint64_t) + sizeof(uint64_t));
 
-        return size;
-    };
+               return size;
+           };
 }
 
 void* ThroughputCommandDataType::createData()
 {
     return (void*)new ThroughputCommandType();
 }
-void ThroughputCommandDataType::deleteData(void* data)
+
+void ThroughputCommandDataType::deleteData(
+        void* data)
 {
     delete((ThroughputCommandType*)data);
 }
